Another fix for Recipe images/cards not deleting

+++ ffdjango/fftracker/fftracker/MealRecipeViews.py
@@ -128,7 +128,7 @@ class RecipeImageView(viewsets.ViewSet):
         queryset = Recipes.objects.filter(r_num = pk)
         img_path = queryset[0].r_img_path
         queryset[0].update(r_img_path = None)
-        if (img_path):
+        if (img_path and os.path.exists('var/www/html' + img_path)):
             os.remove('var/www/html/' + img_path)
         return Response(200)
 
@@ -160,7 +160,7 @@ class RecipeCardView(viewsets.ViewSet):
         queryset = Recipes.objects.filter(r_num = pk)
         card_path = queryset[0].r_card_path
         queryset[0].update(r_card_path = None)
-        if (card_path):
+        if (card_path and os.path.exists('var/www/html' + card_path)):
             os.remove('var/www/html/' + card_path)
         return Response(200)
